+++ web-ui/src/main/resources/catalog/components/elasticsearch/EsFacet.js
@@ -47,6 +47,8 @@
           "resourceType",
           "resourceTitle*",
           "resourceAbstract*",
+          "resourceDate",
+          "resourceEdition",
           "draft",
           "owner",
           "link",